nominax/requirements.txt
Django==1.11.2
django-cuser==2017.3.16
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
supra==1.1.55
